% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{Characterizes the fire regime and meteorology}
\usage{
build_report(
  study.area,
  my.fires,
  my.dated.fires,
  meteo.data,
  active.period,
  user.period,
  meteo.aggregation,
  min.size,
  min.overlap,
  output.folder,
  fire.aggregation,
  fire.size.intervals,
  manual.dur,
  create.clusters
)
}
\arguments{
\item{study.area}{Shapefile with the limits of the study area (polygon). Must not contain more than one polygon.}

\item{my.fires}{Polygon shapefile containing the fire perimeters (dated or not). This shapefile will be used to characterize the fire regime and to propose the durations to use in the MTT simulation. Must contain the fields “Year”, "ID" and "Area_ha".}

\item{my.dated.fires}{Polygon shapefile containing the dated fire perimeters. Same as the dated fire perimeters used in function get.fire.weather}

\item{meteo.data}{A csv file with the meteorological variables in the days with fire events. This file is generated by the function get.fire.weather, and stored as fire_weather_study_area.csv. The user might use a different file, as long as maintaining the same structure.}

\item{active.period}{The most common active period of fire spread. Will be used to select the interval of meteorological values to be used in the creation of clusters or percentiles.}

\item{user.period}{Optional. Numerical vector with the hours to be used in the fire weather characterization.}

\item{meteo.aggregation}{How hourly data should be combined into daily data. Possible methods include: mean (computes the mean of each variable for the active.period), max.min (computes the maximum value of temperature and wind speed and the minimum relative humidity for the active.period), none (does not combine hourly data into daily data and uses all the data in the creation of clusters or percentiles).}

\item{min.size}{Numerical. Minimum fire size to be considered in the analysis. Useful to subset the fire sample. Must be the same as the lowest value set in fire.size.intervals}

\item{min.overlap}{Numerical. Minimal overlap (as percentage) between the fire perimeters and the study area to consider the fire perimeter in the analysis. Must range between 0 and 100.}

\item{output.folder}{Path to the folder where the outputs should be saved.}

\item{fire.aggregation}{How the meteorology of different points inside the same fire perimeter should be combined. Possible methods include: WS (wind speed; the point falling inside the fire perimeter with highest wind speed is kept and the other ignored), none (uses all the points falling inside the fire perimeter).}

\item{fire.size.intervals}{Numeric vector with the fire size classes to be considered in the characterization of the study area.}

\item{manual.dur}{Optional. Numeric vector with the lower limits of each fire size class to be considered as a duration interval. To be used in the MTT calibration process (Run.fcontmtt function)}

\item{create.clusters}{Logical. True returns clusters generated by using kmeans and model-based clustering. False returns percentiles of the meteorological data.}
}
\value{
Returns fuel moisture file based on clustering or percentiles, a csv file with the meteorological values per clustering method, a csv file with a summary of meteorology per fire event, and a csv file with the fire size for the fires recorded in the study area in the period analysed.
}
\description{
Creates files with meteorological characterization of the study area during fire events (csv and fms files); and a word file with the characterization of the fire regime.
}
\examples{
\dontrun{build_report(study.area="C:/user/study_area.shp",
my.fires="C:/user/my_fires.shp",my.dated.fires="C:/user/my_dated_fires.shp",
meteo.data="fire_weather_study_area.csv",active.period="energy",
user.period=c(14,15,16),meteo.aggregation="max.min",
min.size=100,min.overlap=50,fire.aggregation="WS",
fire.size.intervals=c(100,250,500,750,1000,2500,5000,10000),
create.clusters=TRUE,output.folder="C:/user/results")}

}
