% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fire_size.R
\name{evaluate_fire_size}
\alias{evaluate_fire_size}
\title{Compares the simulated fire size distribution with the historical fire size distribution}
\usage{
evaluate_fire_size(
  Folder.Outputs,
  intervals,
  all.dist,
  hist.fire.sizes,
  freqs.durclass,
  plot.all
)
}
\arguments{
\item{Folder.Outputs}{Path to the folder containing the outputs of the FConstMTT runs.}

\item{intervals}{Numerical vector with the intervals of fire size to be considered in the comparison between simulated and the historical fire size distribution.}

\item{all.dist}{Logical. If true, then all the fire size distribution is used to calculate the RMSE and the correlation. If false, then the fire size distribution will only be considered starting from the first numeric value identified in intervals.}

\item{hist.fire.sizes}{Text file (csv) containing the historical fire size. If the function get_fire_weather was used in the process, then the file to be used should be “summary fire size.csv”.}

\item{freqs.durclass}{Text file (csv) with the relative frequency of each meteorological and fuel map scenario used. If the function Gen_ign was used in the process, then the file to be used here should be “clusters_freqs_final.csv”, which is located in the ignition folder.}

\item{plot.all}{Logical. If TRUE, then the fire size distribution of all combinations will be plotted and exported in multiple png files. If FALSE, only the combination with the lowest RMSE is plotted and exported in a single png file.}
}
\value{
Returns figures showing the historical and simulated fire size distribution for all the combinations. Also saves a csv file containing the RMSE and person correlation of each combination of durations and the historical fire size.
}
\description{
Combines and calculates the RMSE and pearson correlation between all the durations simulated and the historical fire size distribution.
}
\examples{
\dontrun{evaluate_fire_size(Folder.Outputs="C:/user/fconstmtt/Outputs",
intervals=c(100,250,500,750,1000,2500,5000,10000),all.dist=FALSE,
hist.fire.sizes="C:/user/summary fire size.csv",
freqs.durclass="C:/user/results/ignitions/clusters_freqs_final.csv")}

}
