% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_weather_nc.R
\name{fire_weather_nc}
\alias{fire_weather_nc}
\title{Get meteorological variables from an existing netcdf file}
\usage{
fire_weather_nc(study.area, my.fires, nc.folder, utc.zone, output.folder)
}
\arguments{
\item{study.area}{Shapefile with the limits of the study area (polygon). Must not contain more than one polygon.}

\item{my.fires}{Polygon shapefile containing the dated fire perimeters. The shapefile must contain a field with a unique id per fire perimeter (ID), a field with the burned area per perimeter in hectares (Area_ha), a field with the date of start of the fire (Date_ini) and the end of the fire (Date_end). The fields Date_ini and Date_end must follow the format yyyy-mm-dd.}

\item{nc.folder}{Path to the folder with the netcdf file.}

\item{utc.zone}{Numeric. Represents the UTC time zone in the study area. Should be set to the time zone during the fire season (i.e. the user should consider the daylight saving time). E.g. for Portugal, the utc.zone should be set to +1.}

\item{output.folder}{Path to the folder where the outputs should be saved.}
}
\value{
Returns a csv file with the hourly meteorological data per fire perimeter in the study area (fire_weather_study_area.csv)
}
\description{
Get meteorological variables from an existing netcdf file and stores it in a csv file.
}
\examples{
\dontrun{fire_weather_nc(study.area="C:/user/study_area.shp",
my.fires="C:/user/my_fires.shp",nc.folder="C:/user/netdcf_file.nc",
output.folder="C:/user/results")}

}
