% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_ign.R
\name{Gen_ign}
\alias{Gen_ign}
\title{Generates ignitions in the study area.}
\usage{
Gen_ign(
  ign.raster,
  IgnitionData,
  fuelmap,
  nIgnitions,
  unburnable,
  LandCover.yr,
  LandCover.wgt,
  shapeOut,
  allow.zero,
  random.ignitions,
  output.folder
)
}
\arguments{
\item{ign.raster}{Raster file with the probability of ignition.}

\item{IgnitionData}{Text file (csv) with the relative frequency of each meteorological cluster (or percentile) and duration to be simulated. If the function build_report was used in the process, then the file to be used should is stored in the folder final_freqs.}

\item{fuelmap}{Fuel map raster file. In alternative to using only one fuel map, the user might specify multiple fuel maps as a vector of characters.}

\item{nIgnitions}{Numeric. Total number of ignitions to be generated.}

\item{unburnable}{Numeric. Value of the fuel model(s) considered unburnable. Ignitions are prevented from being sampled in the fuel model(s) indicated.}

\item{LandCover.yr}{Numeric. Value (or vector of values) of the years of the fuel maps given as input in fuelmap.}

\item{LandCover.wgt}{Numeric. Value (or vector of values) of the weights of each fuel map in the fuelmap and LandCover.yr. The sum of the values must be 1.}

\item{shapeOut}{Binary. If 1, then a point shapefile is generated. If 0, only a text file with the coordinates of the ignitions is generated.}

\item{allow.zero}{Logical. If true, then a 0 percent probability of ignition is allowed in the raster file ign.raster. If false, then a minimum probability corresponding to the 5th percentile of the ign.raster is set. Default is false.}

\item{random.ignitions}{Logical. If true, then ignitions are generated randomly across the study area (excluding in the fuel model identified as unburnable). Default is false.}

\item{output.folder}{Path to the folder where the outputs should be saved.}
}
\value{
Returns text files with the coordinates of the ignitions in the correct format to be used in FConstMTT. One text file is generated by a combination of meteorological cluster, wind direction and duration class. A point shapefile per combination can also be generated.
}
\description{
Generates ignitions in the study area that either follow a surface of probability of ignition (raster file) or are randomly allocated.
}
\examples{
\dontrun{Gen_ign(ign.raster="C:/user/ignition.tif",
IgnitionData="C:/user/results/final_freqs/kmean_3_clusters_final_freqs.csv",
fuelmap=c("C:/user/fuelmap_2000.asc", "C:/user/fuelmap_2010.asc"),
nIgnitions=5000,
unburnable=c(98),
LandCover.yr=c(2000,2010),
LandCover.wgt=c(0.4,0.6),
shapeOut=1,
allow.zero=FALSE,
random.ignitions=FALSE,
output.folder="C:/user/results")}
}
