% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fire_weather.R
\name{get_fire_weather}
\alias{get_fire_weather}
\title{Downloads ERA5-Land reanalysis data and stores it in a csv file}
\usage{
get_fire_weather(
  study.area,
  my.fires,
  output.folder,
  utc.zone,
  wf_user,
  wf_key
)
}
\arguments{
\item{study.area}{Shapefile with the limits of the study area (polygon). Must not contain more than one polygon.}

\item{my.fires}{Polygon shapefile containing the dated fire perimeters. The shapefile must contain a field with a unique id per fire perimeter (ID), a field with the burned area per perimeter in hectares (Area_ha), a field with the date of start of the fire (Date_ini) and the end of the fire (Date_end). The fields Date_ini and Date_end must follow the format yyyy-mm-dd.}

\item{output.folder}{Path to the folder where the outputs should be saved.}

\item{utc.zone}{Numeric. Represents the UTC time zone in the study area. Should be set to the time zone during the fire season (i.e. the user should consider the daylight saving time). E.g. for Portugal, the utc.zone should be set to +1.}

\item{wf_user}{Personal UID that identifies the user in the Climate Change Service (CDS). Can be retrieved from the user profile page after logging in at https://cds.climate.copernicus.eu/user}

\item{wf_key}{Personal API Key that identifies the user in the Climate Change Service (CDS). Can be retrieved from the user profile page after logging in at https://cds.climate.copernicus.eu/user}
}
\value{
Returns a raw netdcf file with all the hourly meteorological variables of temperature, relative humidity, wind speed and direction (same as simply using the wf_request function from the ecmwfr package; the request is stored at https://cds.climate.copernicus.eu/cdsapp#!/yourrequests). Also returns a csv file with the hourly meteorological data per fire perimeter in the study area (fire_weather_study_area.csv)
}
\description{
Downloads ERA5-Land reanalysis data and stores it in a csv file
}
\examples{
\dontrun{get_fire_weather(study.area="C:/user/study_area.shp", my.fires="C:/user/my_fires.shp",
utc.zone=+1,wf_user="12345",wf_key="123456ab-12a3-1234-1234-123ab4567891",
output.folder="C:/user/results")}

}
