% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_number_ign.R
\name{get_number_ign}
\alias{get_number_ign}
\title{Estimates the minimum number of ignitions that should be used for the calibration process}
\usage{
get_number_ign(fuelmap, unburnable)
}
\arguments{
\item{fuelmap}{Fuel map raster file.}

\item{unburnable}{Numeric. Value of the fuel model(s) considered unburnable.}
}
\value{
Returns the value of the minimum number of ignitions to use in the calibration process.
}
\description{
Estimates the minimum number of ignitions that should be used for the calibration process
}
\examples{
\dontrun{get_number_ign(fuelmap="C:/fuel_model_map.asc",
unburnable=c(97,98))}

}
