% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fconstmtt.R
\name{run_fconstmtt}
\alias{run_fconstmtt}
\title{Create the inputs and runs FConstMTT from batch files}
\usage{
run_fconstmtt(
  Folder,
  landscape,
  CrownFireMethod,
  customfmd,
  FmsFolder,
  Resolution,
  GridDistanceUnits,
  MeteoFile,
  Duration.1,
  Duration.2,
  Duration.3,
  Duration.4,
  Duration.5,
  WS_unit,
  FireListFile,
  SpotProbability,
  output.folder,
  OutputFirePerims,
  MetricFLP,
  Run.fconstmtt
)
}
\arguments{
\item{Folder}{Path to the folder containing the bin folder with the FContMTT executable. This will also be the folder where the input and the batch files will be saved.}

\item{landscape}{Path to the folder contained the landscape file(s) to be used (either .tif or .lcp)}

\item{CrownFireMethod}{Numerical. Either 0 for the Finney's method, or 1 for use of Scott and Rheinhardt's method.}

\item{customfmd}{File containing the characterization of custom fuel models (.fmd file).}

\item{FmsFolder}{Path to the folder containing the fms file(s) to be use in the simulation.}

\item{Resolution}{Numeric. Represents the desired resolution of calculations.}

\item{GridDistanceUnits}{Numeric. Sets grid distance units, where 0 is for meters and 1 is for feet.}

\item{MeteoFile}{Text file containing the meteorological characterization of the clusters or percentiles to be used in the simulation.}

\item{Duration.1}{Vector of values that define the interval of duration values to be tested in the first duration class. The vector must contain three values: c(from, to, by).}

\item{Duration.2}{Optional. vector of values that define the interval of duration values to be tested in the second duration class. The vector must contain three values: c(from, to, by).}

\item{Duration.3}{Optional. vector of values that define the interval of duration values to be tested in the third duration class. The vector must contain three values: c(from, to, by).}

\item{Duration.4}{Optional. vector of values that define the interval of duration values to be tested in the fourth duration class. The vector must contain three values: c(from, to, by).}

\item{Duration.5}{Optional. vector of values that define the interval of duration values to be tested in the fifth duration class. The vector must contain three values: c(from, to, by).}

\item{WS_unit}{Either kmh or mph.}

\item{FireListFile}{Path to the folder containing the ignitions to be used.}

\item{SpotProbability}{Numerical. Value to be used in the spot probability during the FConstMTT simulations.}

\item{output.folder}{Path to the folder where the outputs should be saved.}

\item{OutputFirePerims}{Binary. If 1, then the simulated fire perimeters are stored. If 0, then the simulated fire perimeters are not stored.}

\item{MetricFLP}{If 1, then the outputs of flame length probability (FLP file) are stored in meters. If 0, then the FLP files are stored in feet.}

\item{Run.fconstmtt}{Binary. If 1, then FConstMTT will automatically start running after creating the required files. If 0, then the .input and batch files are created but FConstMTT does not start running.}
}
\value{
Returns the input and batch files required to run FConstMTT for all the combinations to be tested. Optionally, FConstMTT might be run from R after the creation of the required files.
}
\description{
Uses the files generated by the functions build_report and Gen_ign to generate the input and batch files and to run FConstMTT.
}
\examples{
\dontrun{run_fconstmtt(Folder="C:/user",landscape="C:/user/landscape_files",
CrownFireMethod=1,customfmd="C/user/my_fmd.fmd",
FmsFolder="C:/user/results/FMS_files/kmeans_3_clusters",
Resolution=100,GridDistanceUnits=0,
MeteoFile="C:/user/results/clusters_meteo/kmeans_3_clusters.csv",
Duration.1=c(min,max,by),
Duration.2=c(min,max,by),
Duration.3=c(min,max,by),
Duration.4=c(min,max,by),
Duration.5=c(min,max,by),
WS_unit="kmh",
FireListFile="C:/user/results/ignitions",SpotProbability=0.05,
output.folder="C:/user/fconstmtt/Outputs",
OutputFirePerims=1,MetricFLP=0,
Run.fconstmtt=1)}

}
