% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_BP_nxburned.R
\name{check_BP_nxburned}
\alias{check_BP_nxburned}
\title{Compares the estimated burn probability with the historical number of times burned.}
\usage{
check_BP_nxburned(
  Folder.Outputs,
  freq.scenario,
  choose.combos,
  combos.file,
  obs.nxburned,
  group.nxburned.from,
  export.plots
)
}
\arguments{
\item{Folder.Outputs}{Path to the folder containing the outputs of the FConstMTT runs.}

\item{freq.scenario}{Text file (csv) with the relative frequency of each meteorological and fuel map scenario used. If the function Gen_ign was used in the process, then the file to be used here should be “clusters_freqs_final.csv”, which is located in the ignition folder.}

\item{choose.combos}{Numerical vector. Specify the combinations to be tested. The numeric identification of the combinations is located in the combos.file. In alternative, select “all” to use all duration combinations or "best" to use only the combination with the lowest RMSE.}

\item{combos.file}{A text file (csv) with the numeric identification of the different combinations. If check_fire_size was run, then this file was stored as “rmse_combos.csv”.}

\item{obs.nxburned}{Raster file with the historical number of times burned. This raster must have the same spatial resolution and alignment that the simulated burn probability rasters.}

\item{group.nxburned.from}{A numeric value representing the maximum value of number of times burned that should be considered in the correlation analysis. A value of five means that cells with five or more number of times burned will be grouped together. Optional}

\item{export.plots}{Binary. If 1, then a boxplot showing the correlation between the estimated burn probability and the number of times burned is saved. If 0, no plot is saved.}
}
\value{
Returns a raster file with the simulated burn probability, a text file (csv) showing the pearson correlation between the simulated burn probability for each combination and the historical number of times burned, and a text file with the relative frequency per class of number of times burned. Optionally, it can also return a boxplot showing the same correlation.
}
\description{
Combines and calculates the pearson correlation between the estimated burn probability that result from multiple duration combinations and the historical number of times burned.
}
\examples{
\dontrun{check_BP_nxburned(Folder.Outputs="C:/user/fconstmtt/Outputs",
freq.scenario="C:/user/results/ignitions/clusters_freqs_final.csv",
choose.combos=c(1,2),combos.file="C:/user/fconstmtt/Outputs/rmse_combos.csv",
group.nxburned.from = 5,
obs.nxburned="C:/user/number_of_times_burned.tif",
export.plots=1)}

}
